+++ src/main/kotlin/com/stark/shoot/domain/service/chatroom/ChatRoomEventService.kt
@@ -19,7 +19,7 @@ class ChatRoomEventService {
      */
     fun createChatRoomCreatedEvents(chatRoom: ChatRoom): List<ChatRoomCreatedEvent> {
         val roomId = chatRoom.id ?: throw IllegalArgumentException("채팅방 ID가 없습니다.")
-        
+
         // 각 참여자에 대한 이벤트 생성
         return chatRoom.participants.map { participantId ->
             ChatRoomCreatedEvent.create(
@@ -28,4 +28,5 @@ class ChatRoomEventService {
             )
         }
     }
+
 }
